pyproject.toml
[tool.poetry]
name = "graphql_fastapi"
version = "0.1.0"
description = ""
authors = ["Lievi Silva <lievi.silva@luizalabs.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
sqlmodel = "^0.0.4"
uvicorn = "^0.15.0"
strawberry-graphql = "^0.90.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.910"
black = {version = "^21.11b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
